refactor(telemetry): address final review findings (public import, ikey validation, fixture dedup)
